// contracts/nameservice/Nameservice.types.ts
/**
* This file was automatically generated by @cosmwasm/ts-codegen@0.24.0.
* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
* and run the @cosmwasm/ts-codegen generate command to regenerate this file.
*/

export type Uint128 = string;
export interface InstantiateMsg {
  purchase_price?: Coin | null;
  transfer_price?: Coin | null;
}
export interface Coin {
  amount: Uint128;
  denom: string;
  [k: string]: unknown;
}
export type ExecuteMsg = {
  register: {
    name: string;
  };
} | {
  transfer: {
    name: string;
    to: string;
  };
};
export type QueryMsg = {
  resolve_record: {
    name: string;
  };
} | {
  config: {};
};
export interface ConfigResponse {
  purchase_price?: Coin | null;
  transfer_price?: Coin | null;
}
export interface ResolveRecordResponse {
  address?: string | null;
}